Renamed 'bridge' namespace to 'executor' to match all of the code in 'src/executor'
