+++ test/executor/copy_test.cpp
@@ -80,7 +80,7 @@ TEST_F(CopyTests, Copying) {
     std::vector<type::Value> params;
     std::vector<int> result_format(statement->GetTupleDescriptor().size(), 0);
     std::vector<StatementResult> result;
-    bridge::ExecuteResult status = traffic_cop.ExecuteStatementPlan(
+    executor::ExecuteResult status = traffic_cop.ExecuteStatementPlan(
         statement->GetPlanTree().get(), params, result, result_format);
     EXPECT_EQ(status.m_result, peloton::ResultType::SUCCESS);
     LOG_TRACE("Statement executed. Result: %s",
