+++ test/statistics/stats_test.cpp
@@ -387,7 +387,7 @@ TEST_F(StatsTests, MultiThreadStatsTest) {
 //  std::vector<type::Value> params;
 //  std::vector<StatementResult> result;
 //  std::vector<int> result_format(statement->GetTupleDescriptor().size(), 0);
-//  bridge::ExecuteResult status = traffic_cop.ExecuteStatementPlan(
+//  executor::ExecuteResult status = traffic_cop.ExecuteStatementPlan(
 //      statement->GetPlanTree().get(), params, result, result_format);
 //  LOG_TRACE("Statement executed. Result: %s",
 //            ResultTypeToString(status.m_result).c_str());
